backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0
anthropic>=0.18.1
python-dotenv>=1.0.1
pydantic>=2.6.0
dateparser>=1.2.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
